+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Fixes
 
 * Fixed `gv` to account for Graphviz HTML-like labels requiring certain characters, namely the set "<>&, to be escaped in Graphviz HTML-like labels, and removed completely in attribute values.
+* Fixed `df_equiv` to properly handle data frames with zero columns or duplicate rows.
 
 # autodb 2.0.0